refactor: extract duplicated pad() and formatDelta() to shared utility
Create src/views/format-utils.ts exporting pad() and formatDelta()
functions that were previously duplicated across four view files:
- src/plan/views/comparison.ts
- src/plan/views/list.ts
- src/implement/views/comparison.ts
- src/implement/views/list.ts

Both utilities handle ANSI escape codes correctly for terminal
formatting:
- pad() strips ANSI codes for accurate width calculations
- formatDelta() formats percentage changes with green/red coloring

This refactoring reduces code duplication by ~80 lines and ensures
consistent formatting behavior across all comparison views.
